requirements.txt
pandas>=1.5.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
requests>=2.28.0
python-dotenv>=0.19.0 